backend/requirements.txt
flask
python-dotenv
flask_restful